requirements.txt
amqp==1.4.6
anyjson==0.3.3
arrow==0.6.0
AST==0.0.2
Babel==2.3.4
BeautifulSoup==3.2.1
beautifulsoup4==4.4.1
billiard==3.3.0.20
boto==2.45.0
celery==3.1.18
certifi==2015.4.28
cffi==1.1.2
cryptography==0.9.3
dj-database-url==0.4.1
Django==1.8.2
django-braces==1.8.0
django-countries==3.3
django-debug-toolbar==1.3.0
django-extensions==1.5.5
django-formtools==1.0
django-generic-positions==0.2.2
django-haystack==2.5.1
django-hvad==1.6.0
django-libs==1.67.6
django-location-field==1.6
django-mptt==0.7.4
django-otp==0.3.8
django-paypal==0.3.3
django-phonenumber-field==0.7.2
django-review==1.9.5
Django-Select2==4.3.1
django-session-security==2.4.0
django-smart-selects==1.1.1
django-sslify==0.2.7
django-storages==1.5.1
django-suit==0.2.14
django-suit-redactor==0.0.2
django-tracking2==0.3.3
django-two-factor-auth==1.4.0
django-user-media==1.2.3
django-uuidfield==0.5.0
django-versatileimagefield==1.6.3
djangorestframework==3.4.6
easy-thumbnails==2.2
elasticsearch==2.3.0
enum34==1.0.4
gunicorn==19.6
httplib2==0.9.2
idna==2.0
ipaddress==1.0.14
ipython==3.2.1
Jinja2==2.7.3
jsonfield==1.0.3
kombu==3.0.26
MarkupSafe==0.23
ndg-httpsclient==0.4.0
nltk==3.2.1
numpy==1.11.2
oauthlib==1.1.2
pdb==0.1
phonenumbers==7.7.5
Pillow==2.8.2
psycopg2==2.6
pyasn1==0.1.8
pycparser==2.14
PyJWT==1.4.0
pyOpenSSL==0.15.1
pysolr==3.6.0
python-dateutil==2.4.2
python-gnupg==0.3.7
python-magic==0.4.6
python-openid==2.2.5
python-social-auth==0.2.19
pytz==2015.4
PyYAML==3.11
qrcode==4.0.4
redis==2.10.3
requests==2.10.0
requests-oauthlib==0.6.1
requests-toolbelt==0.7.0
sendgrid-django==4.0.4
simplejson==3.8.2
six==1.10.0
sqlparse==0.1.16
twilio==5.6.0
unicode-slugify==0.1.3
Unidecode==0.4.18
wheel==0.24.0
winappdbg==1.5
blockchain==1.3.3
